+++ pkg/cluster/retry/noretry/no_retry.go
@@ -6,7 +6,7 @@
  * (the "License"); you may not use this file except in compliance with
  * the License.  You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -15,11 +15,7 @@
  * limitations under the License.
  */
 
-package count_based
-
-import (
-	"fmt"
-)
+package noretry
 
 import (
 	"github.com/apache/dubbo-go-pixiu/pkg/cluster/retry"
@@ -27,37 +23,25 @@ import (
 )
 
 func init() {
-	retry.RegisterRetryPolicy(model.RetryerCountBased, newCountBasedRetry)
+	retry.RegisterRetryPolicy(model.RetryerNoRetry, newNoRetryPolicy)
 }
 
-type CountBasedRetry struct {
-	MaxAttempts uint
-	currentTry  uint
+type NoRetryPolicy struct {
+	firstTime bool
 }
 
-func (r *CountBasedRetry) Attempt(err error) bool {
-	if r.currentTry < r.MaxAttempts {
-		r.currentTry++
-		return true
+func (n *NoRetryPolicy) Attempt(err error) bool {
+	if !n.firstTime {
+		n.firstTime = true
+		return true // Allow the first attempt
 	}
 	return false
 }
 
-func (r *CountBasedRetry) Reset() {
-	r.currentTry = 0
+func (n *NoRetryPolicy) Reset() {
+	n.firstTime = false
 }
 
-func newCountBasedRetry(config map[string]any) (retry.Retryer, error) {
-	timesValue, exists := config["times"]
-	if !exists {
-		return nil, fmt.Errorf("'times' field is missing in retry configuration")
-	}
-
-	timesUint, ok := timesValue.(int)
-	if !ok {
-		return nil, fmt.Errorf("invalid type for 'retry.count_based.times', expected int but got %T", timesValue)
-	}
-
-	// Total attempts = 1 initial try plus number of retries.
-	return &CountBasedRetry{MaxAttempts: uint(timesUint) + 1}, nil
+func newNoRetryPolicy(config map[string]any) (retry.Retryer, error) {
+	return &NoRetryPolicy{firstTime: false}, nil
 }
