+++ pkg/cluster/retry/noretry/no_retry_test.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package no_retry
+package noretry
 
 import (
 	"errors"
