+++ pkg/pluginregistry/registry.go
@@ -25,9 +25,9 @@ import (
 	_ "github.com/apache/dubbo-go-pixiu/pkg/cluster/loadbalancer/ringhash"
 	_ "github.com/apache/dubbo-go-pixiu/pkg/cluster/loadbalancer/roundrobin"
 	_ "github.com/apache/dubbo-go-pixiu/pkg/cluster/loadbalancer/weightrandom"
-	_ "github.com/apache/dubbo-go-pixiu/pkg/cluster/retry/count_based"
-	_ "github.com/apache/dubbo-go-pixiu/pkg/cluster/retry/exponential_backoff"
-	_ "github.com/apache/dubbo-go-pixiu/pkg/cluster/retry/no_retry"
+	_ "github.com/apache/dubbo-go-pixiu/pkg/cluster/retry/countbased"
+	_ "github.com/apache/dubbo-go-pixiu/pkg/cluster/retry/exponentialbackoff"
+	_ "github.com/apache/dubbo-go-pixiu/pkg/cluster/retry/noretry"
 	_ "github.com/apache/dubbo-go-pixiu/pkg/filter/accesslog"
 	_ "github.com/apache/dubbo-go-pixiu/pkg/filter/auth/jwt"
 	_ "github.com/apache/dubbo-go-pixiu/pkg/filter/authority"
